Fixed small compiler warning on powerpc-darwin

+++ libraries/clib/io/AudioPlayer.c
@@ -24,7 +24,7 @@ kaffe_applet_AudioPlayer_playFile(struct Hjava_lang_String* jstr)
 	char    fName[MAXPATHLEN];
 	int     bLen = 1024;
 	int     fin, dev, rc;
-	size_t	bRead;
+	ssize_t	bRead;
 	void	*buf;
 
 	buf = KMALLOC( bLen);
@@ -46,7 +46,7 @@ kaffe_applet_AudioPlayer_playFile(struct Hjava_lang_String* jstr)
 		SignalError("java.io.IOException", SYS_ERROR(rc));
 	}
 
-	while ( (KREAD( fin, buf, bLen, &bRead ) == 0) && ((ssize_t)bRead > 0)) {
+	while ( (KREAD( fin, buf, bLen, &bRead ) == 0) && (bRead > 0)) {
 		ssize_t bWritten;
 		KWRITE( dev, buf, bRead, &bWritten );	/* XXX check error */
 	}
